Use a CouchDB view to fetch the list of items
Instead of getting _all_docs, use a view that sorts items by date.
This fixes a bug where design docs in the database would show up as empty items in the list.
It also removes the necessity of constructing document IDs that sort by date; I've left the code in for now because it doesn't cause any problems.
NOTE: There are still some sort problems because some platform isn't writing dates in the correct format; entries created by that will stick to the top until deleted.